added average and mode statistics

+++ cpp/src/stl_util/stl_typedefs.hpp
@@ -3,4 +3,4 @@
 #include <map>
 
 template<typename T>
-using SparseVector = std::map<size_t, T>;
+using SparseVector = std::map<int, T>;
